requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.1
python-dotenv>=1.0.0
numpy>=1.21.0